+++ secator/tasks/feroxbuster.py
@@ -7,7 +7,7 @@ from secator.definitions import (CONTENT_TYPE, DATA, DELAY, DEPTH, FILTER_CODES,
 							   OPT_NOT_SUPPORTED, OPT_PIPE_INPUT, PROXY,
 							   RATE_LIMIT, RETRIES, STATUS_CODE,
 							   THREADS, TIMEOUT, USER_AGENT, WORDLIST, WORDS, URL)
-from secator.output_types import Url
+from secator.output_types import Url, Info
 from secator.serializers import JSONSerializer
 from secator.tasks._categories import HttpFuzzer
 
@@ -15,7 +15,7 @@ from secator.tasks._categories import HttpFuzzer
 @task()
 class feroxbuster(HttpFuzzer):
 	"""Simple, fast, recursive content discovery tool written in Rust"""
-	cmd = 'feroxbuster --auto-bail --no-state'
+	cmd = 'feroxbuster --no-state'
 	input_types = [URL]
 	output_types = [Url]
 	tags = ['url', 'fuzz']
@@ -25,7 +25,8 @@ class feroxbuster(HttpFuzzer):
 	json_flag = '--silent --json'
 	opt_prefix = '--'
 	opts = {
-		# 'auto_tune': {'is_flag': True, 'default': False, 'help': 'Automatically lower scan rate when too many errors'},
+		'auto_bail': {'is_flag': True, 'default': False, 'help': 'Automatically bail out when too many errors occur'},
+		'auto_tune': {'is_flag': True, 'default': True, 'help': 'Automatically lower scan rate when too many errors'},
 		'extract_links': {'is_flag': True, 'default': False, 'help': 'Extract links from response body'},
 		'collect_backups': {'is_flag': True, 'default': False, 'help': 'Request likely backup exts for urls'},
 		'collect_extensions': {'is_flag': True, 'default': False, 'help': 'Discover exts and add to --extensions'},
@@ -53,7 +54,8 @@ class feroxbuster(HttpFuzzer):
 		TIMEOUT: 'timeout',
 		USER_AGENT: 'user-agent',
 		WORDLIST: 'wordlist',
-		'request_headers': 'headers'
+		'request_headers': 'headers',
+		'auto_tune': 'auto-tune',
 	}
 	item_loaders = [JSONSerializer()]
 	output_map = {
@@ -77,6 +79,15 @@ class feroxbuster(HttpFuzzer):
 	proxy_http = True
 
 	@staticmethod
+	def on_cmd(self):
+		rate_limit = self.get_opt_value('rate_limit')
+		auto_tune = self.get_opt_value('auto_tune')
+		if rate_limit is not None and auto_tune:
+			self.add_result(Info(message='Disabling auto-tune since it conflicts with rate-limit'))
+			self.cmd = self.cmd.replace('--auto-tune', '')
+		return self.cmd
+
+	@staticmethod
 	def on_start(self):
 		if self.inputs_path:
 			self.cmd += ' --stdin'
@@ -88,6 +99,13 @@ class feroxbuster(HttpFuzzer):
 		return True
 
 	@staticmethod
-	def on_item(self, item):
-		item.request_headers = self.get_opt_value('header', preprocess=True)
-		return item
+	def on_json_loaded(self, item):
+		yield Url(
+			url=item['url'],
+			method=item['method'],
+			status_code=item['status'],
+			time=item['timestamp'],
+			response_headers=item['headers'],
+			request_headers=self.get_opt_value('header', preprocess=True),
+			confidence='low'
+		)